Boots up to fstabinit().

+++ Kernel/platform-msp430fr5969/config.h
@@ -46,7 +46,7 @@ extern unsigned char vt_mangle_6847(unsigned char c);
 extern int __user_base;
 
 #define TICKSPERSEC 100   /* Ticks per second */
-#define PROGBASE    ((uint16_t)&__user_base)  /* also data base */
+#define PROGBASE    ((uint16_t)(size_t)&__user_base)  /* also data base */
 #define PROGLOAD    PROGBASE /* also data base */
 #define PROGTOP     0xff80  /* Top of program */
