+++ Kernel/usermem.c
@@ -18,8 +18,8 @@ usize_t valaddr(const char *base, usize_t size)
 	   emulation alone can touch below 0x100 */
 	if (!base || base < (const char *)PROGBASE || base + size < base)
 		size = 0;
-	else if (base + size > (const char *)udata.u_top)
-		size = (char *)udata.u_top - base;
+	else if (base + size > (const char *)(size_t)udata.u_top)
+		size = (char *)(size_t)udata.u_top - base;
 	if (size == 0)
 		udata.u_error = EFAULT;
 	return size;
